Fix SIGSEGV: open_tcp() truncated pointer user param to int on LP64
open_tcp() declared its last parameter as 'int user' and stored it into
the 'long' field tcb->user. On LP64 systems a 64-bit pointer passed as
(long)tn was truncated to 32 bits on the way in, so the telnet upcalls
(tn_tx, rcv_char) dereferenced a garbage address and crashed with SIGSEGV.
Widen the parameter to 'long' to match the field.

+++ src/tcp.h
@@ -312,7 +312,7 @@ struct tcb *open_tcp(struct socket *lsocket,struct socket *fsocket,
 	void (*r_upcall)(struct tcb *tcb,int32 cnt),
 	void (*t_upcall)(struct tcb *tcb,int32 cnt),
 	void (*s_upcall)(struct tcb *tcb,enum tcp_state old,enum tcp_state new),
-	int tos,int user);
+	int tos,long user);
 int32 recv_tcp(struct tcb *tcb,struct mbuf **bpp,int32 cnt);
 void reset_all(void);
 void reset_tcp(struct tcb *tcb);
